requirements.txt
playwright~=1.50.0
pytest~=8.3.4
pillow~=11.1.0